Implements mass_assignment_strict_mode
The setting is per class and configures strict handling of unknown attributes
by raising an exception, just like Rails/ActiveModel does.

By default it is disabled, and you need to opt-in.

This is valuable for cases where Granite is used in combination with GQL
defined mutations and there is no risk of using incorrect attributes or
getting some accidental junk from web forms.